Add scoring system interfaces and implementations
This commit introduces a comprehensive scoring system for the bowling tournament management application. Key changes include:

- Added `IView` interface in IScoresView.cs for error handling and data binding.
- Implemented `Adapter` class in RetrieveScoresAdapter.cs for score retrieval with error management.
- Created `ViewModel` and `IViewModel` in ScoreViewModel.cs to represent squad scores.
- Introduced `ScoresExtensions` for dependency injection of the scoring module.
- Added `GridViewModel` and `IGridViewModel` in ScoresGridViewModel.cs for displaying scores and lane assignments.
- Updated `Presenter` class in ScoresPresenter.cs to manage score loading and error handling.
- Created `IView` interface in IUpdateScoresView.cs for score updates.
- Implemented `UpdateScoresAdapter` for handling score updates and error management.
- Added `UpdateScoresPresenter` to manage the score update process.
- Integrated the scoring module into the service collection in PresentationExtensions.cs.
