--- source/main/Scores/Update/IUpdateScoresView.cs
@@ -1,13 +1,0 @@
-﻿namespace BowlingMegabucks.TournamentManager.Scores.Update;
-internal interface IView
-{
-    IEnumerable<IViewModel> Scores { get; }
-
-    void DisplayError(string message);
-
-    void DisplayMessage(string message);
-
-    void KeepOpen();
-
-    void OkToClose();
-}
