--- source/main/Scores/Update/UpdateScoresPresenter.cs
@@ -1,45 +1,0 @@
-﻿
-using Microsoft.Extensions.DependencyInjection;
-
-namespace BowlingMegabucks.TournamentManager.Scores.Update;
-
-internal class Presenter
-{
-    private readonly IView _view;
-
-    private readonly IAdapter _adapter;
-
-    public Presenter(IView view, IServiceProvider services)
-    {
-        _view = view;
-
-        _adapter = services.GetRequiredService<IAdapter>();
-    }
-
-    /// <summary>
-    /// Unit Test Constructor
-    /// </summary>
-    /// <param name="mockView"></param>
-    /// <param name="mockAdapter"></param>
-    internal Presenter(IView mockView, IAdapter mockAdapter)
-    {
-        _view = mockView;
-        _adapter = mockAdapter;
-    }
-
-    public async Task ExecuteAsync(CancellationToken cancellationToken)
-    {
-        await _adapter.ExecuteAsync(_view.Scores, cancellationToken).ConfigureAwait(true);
-
-        if (_adapter.Errors.Any())
-        {
-            _view.DisplayError(string.Join(Environment.NewLine, _adapter.Errors.Select(error => error.Message)));
-            _view.KeepOpen();
-        }
-        else
-        {
-            _view.DisplayMessage("Scores updated");
-            _view.OkToClose();
-        }
-    }
-}
