+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Scores/Retrieve/RetrieveScoresAdapter.cs
