+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Scores/ScoresExtensions.cs
@@ -6,8 +6,9 @@ internal static class ScoresExtensions
 {
     public static IServiceCollection AddScoresModule(this IServiceCollection services)
     {
-        services.AddTransient<RecapSheetForm>();
-        services.AddTransient<Form>();
+        services.AddTransient<Retrieve.IAdapter, Retrieve.Adapter>();
+
+        services.AddTransient<Update.IAdapter, Update.Adapter>();
 
         return services;
     }
