+++ source/presentation/BowlingMegabucks.TournamentManager.Presentation/Scores/Update/UpdateScoresAdapter.cs
